Update BunyanProvider
The BunyanProvider wraps a singleton logger instance. However this
means that if a bit of code sets the log level for itself, it actaully
sets it for everything else that gets a logger from the provider.

This change updates the approach. There is still a singleton root logger,
but calling `getLogger` now returns a custom configured child logger.
Each child adds a `source` field that can be set to any arbitrary string
but should identify the source of the message in some way, e.g. the
class name. The child can also set the log level independently without
interfering with other loggers.